Updated namespace to "SilverStripe"
A bit of a yak shave: We want to plan out work required to incorporate GraphQL in SS4,
to get an idea of the size and impact on our roadmap. Which means we need to get it
on our team board via Zenhub. Which means it needs to be in the 'silverstripe' Github org.
At which point it made sense to switch the PHP namespace to "SilverStripe" as well.