NEW Validate schema keys
Prevents "forgetting" a level. In this example, I've forgotten the "models" key, which means the schema build will just finish without any output or error.

SilverStripe\GraphQL\Schema\Schema:
  schemas:
    default:
      SilverStripe\CMS\Model\SiteTree:
        fields:
          '*': true

The downside of this approach is that we can't easily extend the schema with new keys
through modules building on silverstripe/graphql, but I'd say we cross that bridge when the need arises.