Refactor config loading in Schema
The schema is used for two purposes:
1. Building: It converts config arrays (retrieved from YAML) to instance state, which is required to call `save()` and persist the schema for later use.
2. Fetching: It loads retrieves an existing persisted schema. This does not require any instance state.

`bootConfig()` was called in the constructor, and implied that it prepares the instance for operation. In fact, it only merges enough (global and local) config to set up a SchemaContext object. It does not validate the schema, or build out the instance state (types, models, etc).. At this point, the instance can be used for *some* operations such as retrieving the underlying GraphQL schema (via `fetch()`). This benefitted the main `Controller` use because other instance state is expensive to validate and create from config (via `applyConfig()`). In order to *actually* get a consistent schema (reflecting the YAML configuration), you also needed to call `loadFromConfig()`.

This change centralises the two step process (bootConfig, loadFromConfig) into a single one with an obvious name: `boot()`.
It also removes the need to store the intermediary merged config state on the instance, which is only required for diagnostic purposes and complicates state management.
Since the config state getter was removed, it also required removal of getSignature() - see https://github.com/silverstripe/silverstripe-graphql/pull/343

Context: https://github.com/silverstripe/silverstripe-graphql/pull/335#discussion_r542213347