Remove Schema->getSignature() (#343)
It's not used anywhere, and is "false advertisement"
since there's a whole lot of instance state which it doesn't cover
(e.g. procedural addType() calls, anything in SchemaContext).
If there was a reason to keep it, we shouldn't return a potentially
huge string as a signature, but rather use a hash.
This would also avoid accidentally disclosing potentially
sensitive (serialised) data in cache keys etc.