NEW Revert generic middleware refactor, readd tests
Changing QueryMiddleware to become a generic Middleware
was well intentioned: It could be used for other purposes outside of the query context,
for example if plugins started supporting middleware.
But in order to achieve this, it weakened the (query specific) parameter
signature into a generic $params argument, which makes it harder
to understand how to write valid middleware for developers.
On balance, that's a more important concern than saving a few
lines of code in writing a hypothetical second middleware implementation in the module.

Also re-added the existing tests from GraphQL v3