Use public/_graphql/ subfolder
This is consistent with public/_resources which is also auto-generated by the CMS.
It enables sysadmins to lock down write access to public/ during runtime, with this potentially granular exception.
And it makes it easier to discover which part of the system is responsible for writing these files.

The change is backwards compatible with existing file locations,
although we probably want to remove that before going stable:
https://github.com/silverstripe/silverstripe-graphql/issues/384

Relies on equivalent PR to silverstripe/admin.

The docs are already stating this new location so this was more of an oversight:
https://docs.silverstripe.org/en/4/developer_guides/graphql/getting_started/building_the_schema/#viewing-the-generated-code